+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.0] - 2023-01-20
 
+### Added
+- `conjur-preflight` now detects whether it is writing standard output to a
+  terminal or to a file/pipe and formats with rich or plain text accordingly.
+  [conjurinc/conjur-preflight#19](https://github.com/conjurinc/conjur-preflight/pull/19)
+
 ### Fixed
 - Previously, the application version was not properly embedded in the final
   executable. Now the application and the reports it produces reflect the